requirements.txt
Django
djangorestframework
drf-spectacular
python-dotenv
django-redis
psycopg2-binary
djangorestframework-simplejwt
